refactor(api/app): Phase 2 — merge auth/identity, inline AuthContext
Pure file reorganisation, no logic change.

- Merge auth/identity/{types,resolve-clerk}.ts into a single
  auth/identity.ts; delete the auth/identity/ subdir.
- Delete auth/context.ts; inline AuthContext into trpc.ts next to
  createTRPCContext. trpc.ts now imports identity from ./auth/identity.

trpc.ts and diagnostics.ts stay at api/app/src/ root with middleware and
the procedure ladder inline in trpc.ts — the draft plan's server/ folder
split was dropped per review.
